python/agents/Supply Chain Risk Intelligence System/requirements.txt
# Google Cloud and ADK dependencies
google-adk>=1.0.0
google-cloud-pubsub>=2.18.0
google-cloud-documentai>=2.20.0
google-genai>=0.3.0

# External API clients
tweepy>=4.14.0
aiohttp>=3.8.0
requests>=2.31.0

# Data processing and geospatial
geojson>=3.0.0
geopy>=2.3.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
asyncio>=3.4.3

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0